checking for column existance before dropping

+++ database/migrations/2021_10_06_161455_drop_dpd_profile_id_from_vpn_profile_groups_table.php
@@ -13,9 +13,11 @@ class DropDpdProfileIdFromVpnProfileGroupsTable extends Migration
      */
     public function up()
     {
-        Schema::table('vpn_profile_groups', function (Blueprint $table) {
-            $table->dropColumn('dpd_profile_id');
-        });
+        if (Schema::hasColumn('vpn_profile_groups', 'dpd_profile_id')) {
+            Schema::table('vpn_profile_groups', function (Blueprint $table) {
+                $table->dropColumn('dpd_profile_id');
+            });
+        }
     }
 
     /**
